separate ts types
